requirements.txt
# Core ML frameworks
torch>=2.0.0
pytorch-lightning>=2.0.0
torchmetrics

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0
networkx>=3.0.0

# Plotting and visualization
matplotlib>=3.5.0

# Experiment tracking
wandb>=0.15.0

# Standard library extensions
tqdm
scikit-learn

# Optional optimizers
# muon  # Uncomment if using the Muon optimizer

# Development and utilities
setuptools
wheel